programmatically create world shader
