+++ release/scripts/startup/abler/lib/materials/__init__.py
@@ -15,3 +15,7 @@
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
 # ##### END GPL LICENSE BLOCK #####
+
+from .handlers import *
+from .object import *
+from .world import *
