+++ release/scripts/startup/abler/lib/shadow.py
@@ -31,13 +31,12 @@ def changeSunRotation(self, context):
     acon_sun.rotation_euler.y = 0
     acon_sun.rotation_euler.z = prop.sun_rotation_z
 
-    world_sun = context.scene.world.node_tree.nodes.get("ACON_node_sunVector")
+    nodes = context.scene.world.node_tree.nodes
+    node_world_surface = nodes.get("ACON_nodeGroup_world_surface")
 
-    if world_sun:
-        world_sun_rotation = world_sun.inputs.get("Rotation").default_value
-        world_sun_rotation.x = acon_sun.rotation_euler.x
-        world_sun_rotation.y = acon_sun.rotation_euler.y
-        world_sun_rotation.z = acon_sun.rotation_euler.z
+    if node_world_surface:
+        input_sun = node_world_surface.inputs.get("Sun Vector")
+        input_sun.default_value = acon_sun.rotation_euler
 
 
 def toggleSun(self, context):
